Add special handling for index.md files
Adds a minor bit of magic for index.md files so they are written to the
root of their build directory instead of a nested "index/" directory.
Also changes template context to include file data for sibling pages
only for "index.md" files.

Creates constants and markdown sub-modules to simplify the main module.