refactor: improve performance; Python 3.11 min (#101)
* refactor: improve performance

Improves performance by making better use of concurrency for extracting
data from Markdown files and processing images

* docs: update dependencies

* breaking: make minimum Python 3.11

Updates the minimum Python version to 3.11 for its built-in tomllib and
removes the toml dependency.

toml wasn't pickling properly, causing issues with multiprocessing

* breaking: change non-schema variable prefix

Changes the non-schema front matter variable prefix from ~ to _ so they
don't have to be quoted. tomllib throws an error for an unquoted key
starting with "~"

* docs: add note about non-schema front matter

Adds a note to the Markdown docs about non-schema front matter.

Also adds a link to the settings file in GitHub to show the default
values