Overhaul schema validation (#121)
Overhauls schema validation to use custom-coded Schema.org types rather than https://github.com/blurry-dev/pydantic2-schemaorg/.

Performance of pydantic2-schemaorg was slow and the code was tricky to maintain. The approach of this PR is to focus on Schema.org types that Google uses for rich search results, which is the main benefit of this feature.

Not all schema types and fields are finished, but this is a good start. See here for the Google docs on their supported schema types:

https://developers.google.com/search/docs/appearance/structured-data/search-gallery